Change rcat provider from pygments to bat
This is a great alternative to pygments for syntax highlighting terminal
output to pygments. It's built for purpose, and seems to be much faster
than the alternative, being written in Rust, rather than Python.

Goodness knows if the code quality is better, it certainly seems like
adding custom syntax highlighting would be harder, but being faster
makes a big difference for me, and writing something in a native
language does make a big difference.